Child functions moved to child.c (#20)
* Child functions moved to child.c


* Added debugging


* Added debugging


* Fixes


* More fixes


---------

+++ tests/funcs.sh
@@ -87,7 +87,7 @@ prepare() {
 
    chmod 644 "$PEMFILE"
 
-   "$MC_COMMAND" cp ibm/piler-ci/kospam.sql.gz .
+   [[ -f kospam.sql.gz ]] || "$MC_COMMAND" cp ibm/piler-ci/kospam.sql.gz .
 
    ls -la
